requirements.txt
# Базовые зависимости
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Работа с голосом и распознаванием
openai-whisper>=20231117
SpeechRecognition>=3.10.0
PyAudio>=0.2.13

# Интерфейс и специфические модули
tkinterdnd2>=0.3.0

# Дополнительно (могут потребоваться для whisper/api)
numpy<2.0.0
openai>=1.0.0